Rewrite ContentWithSidebar
